add support for background-size: cover
To determine the dimensions of the image when the background-size is set
to cover, we first compare the aspect ratio (width/height) of the image
with that of the element. If the image's aspect ratio is greater than
the element's, we know that it is fatter than the element and so the
scale would be set by the ratio of their heights. The opposite is true
if the image's aspect ratio is less than that of the element.

Note that there 'cover' is just one of the many ways we can specify
background-size and we haven't taken into account of the others.