refactor: define plugin class
This refactoring will ultimately allow us to call methods on elements
that had draggle background enabled. For example, we could add a method
to disable the draggble background feature bound to the element.

For this, we define the Plugin class and save an instance of it onto the
element's data attribute provided by jQuery.