prevent child elements from triggering draggable background
If a child of the element with draggable background is called, we will
end up calling `e.preventDefault()`. This is problematic if the child
element is a form element which `preventDefault()` will effectively make
unfocusable.

Check whether the clicked element is the actual element with the
draggable background and not a child element before enabling dragging.